requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
streamlit
python-docx
python-dotenv
PyPDF2
numpy
pytesseract
Pillow
google-generativeai
langgraph
pandas
email-validator
python-dateutil
docx2txt
reportlab
pypdf